pyproject.toml
[tool.poetry]
name = "bac-website"
version = "0.1.0"
description = ""
authors = ["MarkOfosu <124046783+MarkOfosu@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.7"
django-filter = "^23.2"
django-heroku = "^0.3.1"
pillow = "^10.0.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
